Added two new USB commands: 0x61 - CMD_GET_CW_TX_FREQ 0x60 - CMD_SET_CW_TX_FREQ These new commands are used to tell the Omnia the CW center frequency up front, so the Omnia will know where to transmit when the operator touches the paddles.
Switched to 4ms CW envelope ramp up / down times, reduced the mute
time to 10ms, moved the CW shaping waveform to RAM to be able
to adjust the IQ balance in the future.

When sending a CW IQ waveform, a slow memcpy has been replaced
by a DMA source pointer juggling. This allows one to keep both
the RX and TX USB connections open while allowing the standalone CW keyer.

Iambic keyer verilog source code has been modified to use a datapath
for an 8-bit counter. This change significantly reduced the UDB usage
and it was necessary for the Iambic Mode A to fit the chip.